pyproject.toml
[tool.poetry]
name = "persona-backend"
version = "0.1.0"
description = "Backend-part of a multi-brand clothing boutique"
authors = ["PERSONA shop"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
djangorestframework = "^3.14.0"
sshtunnel = "^0.4.0"
requests = "^2.29.0"
autopep8 = "^2.0.2"
django-mptt = "^0.14.0"
djangorestframework-simplejwt = "^5.2.2"
djangorestframework-jwt = "^1.11.0"
transliterate = "^1.10.2"
pillow = "^9.5.0"
django-enumfields = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"